move helper functions to lib
